task: added abstractions, projects structure changed
