+++ jobs/Backend/Task/ExchangeRate.Contracts/Models/ExchangeRate.cs
@@ -1,0 +1,22 @@
+﻿namespace ExchangeRate.Contracts.Models;
+
+public record class ExchangeRate
+{
+    public ExchangeRate(Currency sourceCurrency, Currency targetCurrency, decimal value)
+    {
+        SourceCurrency = sourceCurrency;
+        TargetCurrency = targetCurrency;
+        Value = value;
+    }
+
+    public Currency SourceCurrency { get; }
+
+    public Currency TargetCurrency { get; }
+
+    public decimal Value { get; }
+
+    public override string ToString()
+    {
+        return $"{SourceCurrency}/{TargetCurrency}={Value}";
+    }
+}
