+++ jobs/Backend/Task/ExchangeRate.Core.Tests/Providers/CnbExchangeRateProviderTests.cs
@@ -1,0 +1,5 @@
+﻿namespace ExchangeRate.Core.Tests.Providers;
+
+internal class CnbExchangeRateProviderTests
+{
+}
