+++ jobs/Backend/Task/ExchangeRate.Core/Providers/CachedExchangeRateProviderBase.cs
@@ -1,0 +1,5 @@
+﻿namespace ExchangeRate.Core.Providers;
+
+public abstract class CachedExchangeRateProviderBase
+{
+}
