+++ jobs/Backend/Task/ExchangeRate.Core/Providers/Interfaces/IExchangeRateProvider.cs
@@ -1,0 +1,8 @@
+﻿using ExchangeRate.Contracts.Models;
+
+namespace ExchangeRate.Core.Providers.Interfaces;
+
+public interface IExchangeRateProvider
+{
+    Task<IEnumerable<Contracts.Models.ExchangeRate>> GetExchangeRatesAsync(IEnumerable<Currency> currencies);
+}
