+++ jobs/Backend/Task/ExchangeRateUpdater.sln
@@ -5,11 +5,11 @@ VisualStudioVersion = 17.6.33829.357
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExchangeRateUpdater", "ExchangeRateUpdater\ExchangeRateUpdater.csproj", "{7B2695D6-D24C-4460-A58E-A10F08550CE0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRate.Providers", "ExchangeRate.Providers\ExchangeRate.Providers.csproj", "{27B29624-4565-4691-8406-99481E8E7158}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExchangeRate.Core", "ExchangeRate.Core\ExchangeRate.Core.csproj", "{27B29624-4565-4691-8406-99481E8E7158}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRate.Contracts", "ExchangeRate.Contracts\ExchangeRate.Contracts.csproj", "{A28801E5-4E13-45B2-8986-C0F769CA7539}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExchangeRate.Contracts", "ExchangeRate.Contracts\ExchangeRate.Contracts.csproj", "{A28801E5-4E13-45B2-8986-C0F769CA7539}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRate.Providers.Tests", "ExchangeRate.Providers.Tests\ExchangeRate.Providers.Tests.csproj", "{075199E0-AE18-4BD0-B517-EACB0DAE571C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExchangeRate.Core.Tests", "ExchangeRate.Core.Tests\ExchangeRate.Core.Tests.csproj", "{075199E0-AE18-4BD0-B517-EACB0DAE571C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
