--- jobs/Backend/Task/ExchangeRateUpdater/Currency.cs
@@ -1,20 +1,0 @@
-﻿namespace ExchangeRateUpdater
-{
-    public class Currency
-    {
-        public Currency(string code)
-        {
-            Code = code;
-        }
-
-        /// <summary>
-        /// Three-letter ISO 4217 code of the currency.
-        /// </summary>
-        public string Code { get; }
-
-        public override string ToString()
-        {
-            return Code;
-        }
-    }
-}
